Moving ScalaMetaMatchers out to scala-support
