+++ modules/scala-http4s/src/test/scala/tests/core/FullyQualifiedNames.scala
@@ -4,7 +4,7 @@ import org.scalatest.funsuite.AnyFunSuite
 import org.scalatest.matchers.should.Matchers
 import scala.meta._
 
-import support.SwaggerSpecRunner
+import support.{ ScalaMetaMatchers, SwaggerSpecRunner }
 
 import dev.guardrail.Context
 import dev.guardrail.generators.ProtocolDefinitions
@@ -12,7 +12,7 @@ import dev.guardrail.generators.scala.http4s.Http4s
 import dev.guardrail.generators.{ Client, Clients }
 import dev.guardrail.terms.protocol.ClassDefinition
 
-class FullyQualifiedNames extends AnyFunSuite with Matchers with SwaggerSpecRunner {
+class FullyQualifiedNames extends AnyFunSuite with Matchers with SwaggerSpecRunner with ScalaMetaMatchers {
 
   val swagger =
     """
