+++ modules/scala-support/src/test/scala/support/ScalaMetaSupport.scala
@@ -1,0 +1,15 @@
+package support
+
+import org.scalactic.Equality
+import scala.meta.Tree
+
+trait ScalaMetaMatchers {
+  implicit def TreeEquality[A <: Tree]: Equality[A] =
+    new Equality[A] {
+      def areEqual(a: A, b: Any): Boolean =
+        b match {
+          case x: Tree => a.structure == x.structure
+          case _       => false
+        }
+    }
+}
